requirements.txt
pandas
numpy
requests
pyyaml
python-dotenv
scikit-learn
joblib
streamlit
matplotlib
seaborn
sqlalchemy
